libdvbv5: reimplement the logic that gets a full section
The logic that gets a full section is broken: it expects to
receive the sections in order, and not to receive the same
section segment twice.

Reimplement it.

As a side effect, it no longer needs to wait for a section
repeat, with speeds up the code.

On my tests with an ATSC signal, before this patch, scanning
to a single frequency was spending 1.850 secs. After the
patch, dvbv5-scan is now spending 1.432 secs.

That means a 418 ms save per frequency.

Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>