v4l2-compliance: support menu controls with more than 32 items
v4l2-compliance would fail if it encounters menu controls with more
than 32 items. This patch improves that behavior by allowing any
number of items, although only up to 64 will be tested.

This is because v4l2-compliance uses a bitmask for the valid menu
items which is now 64 bits.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>