v4l2-compliance: add helper headers
In order to simplify working with single and multiple planar formats
these two helper headers were created: v4l-helpers.h is the C version
and cv4l-helpers.h is a C++ version on top of v4l-helpers.h.

For now this will only be used in v4l2-compliance, but the intention
is to use it elsewhere as well eventually.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>