qv4l2: add support for V4L2_PIX_FMT_NV16M/61M
Add support for the NV16M and NV61M planar YUV 4:2:2 formats.

This is not the prettiest code. Once the (c)v4l-helpers.h are more
mature (particularly w.r.t. the v4l2-format handling) the qv4l2 code
will switch to that and that should greatly improve the code.

For now though it is good to have support for at least one multiplanar
format in qv4l2 to help testing the multiplanar API.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>