v4l2-compliance: M2M cap + input or output caps should be an error.
Currently this is a warning, but this is now an obsolete combination
and should never be used.

In the very beginning when m2m devices were introduced they were marked
as capture+output devices, but some applications scan video devices for
those that have the CAPTURE cap set (quite reasonable), and they would
also match such m2m devices. Quite soon we realized that this was a
problem and we introduced the m2m cap.

Because of this the combination of M2M + CAPTURE or OUTPUT caps is
illegal and should not be used anymore.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>