qv4l2: add extensive colorspace support
Support for colorspace handling has been added to the qv4l2 tool.
Two new comboboxes have been added: one to select the colorspace
of the captured video (by default it will use the colorspace as
returned by the driver), and one to select the colorspace of the
display which is sRGB by default.

Special care has been taken to do the correct Y'CbCr to R'G'B'
conversion, the R'G'B' to RGB conversion, the actual colorspace
conversion and the conversion to the display colorspace.

Note that colorspace handling is only supported with the openGL
acceleration. Doing this in the software implementation would be
much too slow.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>