qv4l2: improve video output support, esp. multiplanar.
There was rudimentary video output support in qv4l2, at least sufficient
to use it to change things like the format.

But this didn't work if the video output driver used the multiplanar API.
Improved qv4l2 to support this.

Note that qv4l2 assumes that you either have video capture or video output.
Memory-to-memory devices are not currently supported.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>