v4l2-ctl: vastly improve the test pattern generator.
Replace the simplistic test pattern generator by a much improved
version that supports many more test patterns, support colorspace
conversions and a host of other features.

This code is based on the test pattern generator of the upcoming
vivi rewrite and excepts for some small changes in vivi-tpg.h the
code is identical to that driver. In the future the sync-with-kernel
make target of v4l-utils will just copy the code from the kernel,
making the kernel vivi driver the master copy.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>