qv4l2: fix square aspect ratio handling and improve Video Aspect Ratio
Overriding the pixel aspect ratio to a square ratio never worked since
it was somehow missing in the switch statement.

The "4x3" video aspect ratio was moved from last to third position in
the Video Aspect Ratio list, before the widescreen formats. That is a
more logical place for it to be.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>