qv4l2: added fullscreen functionality for capture window
One can enter fullscreen mode by a push button, the F key
or by a double click. To go back to window mode; all
the same options plus the escape key.

An eventfilter is added to show a status bar when moving
the mouse to the bottom of the screen while in fullscreen mode.

Signed-off-by: Anton Arbring <aarbring@cisco.com>
Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>