qv4l2: bypass libv4l2 when obtaining the colorspace
libv4l2 does not update the colorspace information when it should.
So just for the purpose of reading the colorspace information we
bypass libv4l2 and do a direct VIDIOC_G_FMT to the driver.

I kept forgetting why I wasn't getting the right colorspace results
and everytime it took me a while before I realized that the colorspace
was wrong.

Signed-off-by: Hans Verkuil <hansverk@cisco.com>