qv4l2: fix layout issues related to stacked widgets
I've been digging into why on my laptop the stacked frames in the Input section
appear shifted to the right.

The reason is that the initial loop in fixWidth() sets the minimumwidth
without looking at what the widget is. After some debugging I found out that it
was setting the minimumwidth for the wFrameSR widget.

So I added checks that it only changes it for comboboxes, spinboxes and sliders.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>