qv4l2: keep the "Controls" text
Don't chop off the "Controls" text in the tab name: it turns out that
leaving that off makes it hard to tell the difference between tabs
that are setup by qv4l2 and tabs that contain controls from the driver.

If someone can come up with a better way of doing this then that would
be welcome, but for now this is the best I can come up with.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>