v4l2-compliance: allow EIO for control ioctls
The uvc driver has to pass on the new control values to the hardware,
and that may not always accept the new value, even though it was
reported as within the valid range. In that case EIO is returned
and v4l2-compliance should accept that as a valid error code. It's
not the uvc driver's fault that there are many devices with crappy
firmware.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>