qv4l2: fix audio looping
The audio loopback never really works well with qv4l2. After comparing it
with the xawtv3 implementation it became clear that the core problem was
that the wrong alsa device was chosen as the capture device.

The alsa handling has been simplified and only 'hw' devices are shown as
this is what the alsa loop code wants. The "Show All Audio Devices" menu
option has been removed.

In addition, a bug was fixed for radio devices: no associated alsa device
was found and the 'Play' button was disabled, so the audio loop was never
started.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>