v4l2-compliance: require G/S_PRIORITY support for m2m devices
M2M devices didn't support G/S_PRIORITY in the past, but after a
recent kernel change they gained prio support. While that was
actually unintended I realized that it does make sense there as well
since it makes it possible for a filehandle to get exclusive access
to the M2M hardware. So drop the exception for M2M devices.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>