libdvbv5: reimplement the spaces filling logic
Coverity doesn't like this construction:

1371		for (i = strlen(hex); i < 49; i++)
CID 1228899 (#1 of 1): Out-of-bounds access (OVERRUN)20. overrun-buffer-arg: Overrunning array spaces of 50 bytes by passing it to a function which accesses it at byte offset 50 using argument 50UL.
1372                        strncat(spaces, " ", sizeof(spaces));
1373                ascii[j] = '\0';
1374                dvb_loginfo("%s%s %s %s", prefix, hex, spaces, ascii);

Neither do I: strncat implementation is complex and have likely
a loop inside it, in order to find the end of the array. So, this
likely have O(n^2) complexity for no good reason.

Reimplement the logic. This should likely make coverty happy too.

Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>