libdvbv5: Add support for opening FE on read only mode
Currently, all frontend opens happen on R/W mode. That
prevents some types of usage where it is desired to just
monitor an interface without touching it, like what femon
does.

So, add a new method to allow doing that.

Ideally, we should also add (or fix/check) support for
O_NONBLOCK mode, but this would require more efforts.

So, for now, add a todo entry about that.

Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>