libdvbv5: Re-add the status name at the stats report
Changeset d3558a8a8e made the status to disappear.

The main issue that it were trying to fix is that the len were
calculated wrong, but the reason is because len is a pointer,
so it should have been incremented as a pointer, with:

	*len += size

This patch fix it properly, and make the status name to reappear.

Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>