libv4l2: Move alignment of dest_fmt resolution to v4l2_set_src_and_dest_format
So that we always do the alignment, this is necessary because there are
several code paths where dest_fmt gets set to a fmt which has not been
passed through libv4lconvert_try_fmt, and thus is not aligned.

Also call v4lconvert_fixup_fmt when aligment has changed the height / width,
so that bytesperline and sizeimage get set correctly.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Gregor Jasny <gjasny@googlemail.com>