dvb-file: Only print SDT warning if are there any channel to be stored
Sometimes, the PMT table may have some descriptors that don't
belong to any program. Those are already skipped, but a warning
message is printed even if no channel programs is actually there.

Change the logic to only print a warning if both the SDR is not
present and there are program channels on it.

That prevents to show those messages on DVB-C, where some transponders
could be used just for data.

Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>