libdvbv5: print ISDB-T frequencies in MHz
This is just a cosmetic change: when displaying the frequency,
print it in MHz. That makes a little easier to read it.

This also makes more sense, since the frequency on this descriptor
is measured in 1/7 MHz steps. So, the value at the descriptor's
content is 7x the value printed.

Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>