libdvbv5: fix a nasty to track double-free bug
When the entries from a channel file have a name on it, and
a new transponder is added, the code at dvb_scan_add_entry()
would be copying the channel names at the new entries,
and the dvb_file_free() would be trying to free the memory.

Also, the new entry would be polluted by the values that got
filled already from a previous entry.

So, create a fresh entry with just the tunning part of the
entry.

Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>