man: move dvb-format-convert.1 out of Doxygen
Using doxygen to document an userspace man page doesn't make
it easier. Instead, it brings extra troubles, as it would create
all man pages under the very same section.

So, move the dvb-fe-tool man page out of doxygen,
adding it as a separate file.

Hopefully, that will make it easier to maintain.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>