libdvbv5, dvbv5-scan: generalize channel duplication check
When a new channel entry is to be added to the channel list,
it is checked if there's already similar entry in the list.
This check was based only on freq, polarity, allowable frequency shift,
so it could not be used for delivery systems that use stream_id
and mulplex multiple TS's (channels) into one frequency.
This patch adds stream_id to the check.

Signed-off-by: Akihiro Tsukada <tskd08@gmail.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>