configure.ac: Fix gconv compilation with 64 bits
The default lib dir for 64bits is lib64. Change the logic that checks
the default directory for gconv in a way that it will compile fine
with 64 bits.

So, instead of using find, test the common places where the library
is written.

Of course, this could always be overriden by a command line option.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>