dvb, keytable: Workaround a glibc bug when handling locale domain
glibc has a weird logic to handle errors for argp_parse:
instead of using "libc" textdomain internally inside the library,
it uses the one defined by textdomain(). Of course, this won't
work, as we don't have translated messages there for the glibc
stuff.

There's a way to override it, by passing argp_domain, but this
breaks translation of other things.

Maybe I'm just too stupid to find a right way of doing that,
but let's workaround it by adding some extra messages inside
the c files.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>