v4l2-compliance: continue testing with some failures
Usually any failure will exit the test. However, with some failures you should
just continue testing. It's still reported, but the test is allowed to finish.

This is related to the uvc driver where certain failures are not easy to resolve.
It's useful if the full test can be run regardless.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>