v4l2-compliance: fix save/restore state
The frequency wasn't saved and restored correctly. Even if the current
input is not a tuner, but another input is, then you can still set the
tuner frequency.

Instead of checking if the current input is a tuner, check if there is
a tuner at all (and ditto for a modulator for outputs).

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>