v4l2-compliance: m2m streaming fixes
When obtaining the current format, explicit set the type otherwise it will
take it from the node which is not what you want for m2m devices.

Also call reqbufs again after calling create_bufs: the number of buffers for
CAPTURE and OUTPUT must be the same, and they no longer are after the call to
create_bufs.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>