libdvbv5: better handle EINTR/EAGAIN
The DVB API may return EINTR or EAGAIN if an ioctl is called
while the DVB is doing certain tasks. So, add a code that will
retry the ioctl on such cases.

We don't need to do that on the statistics part, as we can simply
assume that the stats is not available on that time (with is true).

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>