v4l2-ctl: fix reduced fps support
I merged the patch adding support for reduced fps a bit too quickly:
if this option is set, then it should set the V4L2_DV_FL_REDUCED_FPS
flag in the v4l2_bt_timings struct. The pixelclock field in that same
struct will remain the nominal pixelclock. The flag is used to signal
to the driver that reduced fps is required, but the driver may clear
it if it cannot slow down the pixelclock by the factor 1000/1001.

Rather than reverting patches, just fix this code (it's just a few
lines).

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>