v4l-utils/contrib/gconv: fix build error with glibc 2.21+
Building this gconv module with glibc > 2.21 failed with errors.

It has copies of glibc's iconv module files,
and #includes them as the base of implemententation,
which is the same impl. pattern as the other internal iconv modules.
But those base files depend on the internal of iconv to some extent,
and it had been changed in glibc 2.21.

Signed-off-by: Akihiro Tsukada <tskd08@gmail.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>