v4l2-compliance: fix prepare_buf tests, drop bytesused == 0 test
The kernel gives a warning when attempting to use bytesused == 0 for
an output buffer. This is confusing when doing the compliance test,
so drop the test where bytesused is set to 0.

A related bug is that while bytesused was set for output buffer when
qbuf was called, this was not done for prepare_buf. Fix this.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>