libmediactl: Don't free an already freed address if no links
There's currently a bug at libmediactl: if the entity returns
0 links, the library crashes with:
	*** Error in `media-ctl': free(): invalid next size (fast): 0x00000000024da1e0 ***
	Aborted

The problem is that the next lines assume that both pads and
links are not zero:
	links.pads = calloc(entity->info.pads, sizeof(struct media_pad_desc));
	links.links = calloc(entity->info.links, sizeof(struct media_link_desc));

While the above won't fail, it will use the links array at the
ioctl() and then it will try to free it.

Fix it by adding a test to check if those integers are zero.
If so, return from the routine doing nothing.

While here, add an extra debug message to allow seeing how many
pads/links an entity is reporting.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>