Revert "libmediactl: add support for the new DVB devnodes"
We ended by discontinuing devnode support at entity, and decided
to use a different mapping for DVB.

Also, the change added at libmediactl is for the current MC DVB
stuff, marked as BROKEN on the Kernel, as we decided in favor of
remanufacturing the entire Media Controller support using a different
approach ("MC next generation"), with makes the changes there
obsolete.

As we're about to release a version 1.8.0 of v4l-utils, revert
this patch.

A latter change, at 1.9.x series, will be adding support for MC
next generation.

Please notice that the changes at libmediactl to get the /dev
from the sysfs information for DVB is independent, and it is OK
with the next gen. So, it won't be reverted.

This reverts commit 59c3c11afef4e4a14fb559529dc4456dec17d2ee.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>