README: add the instructions about how to build v4l-utils
Even for me, installing v4l-utils on some new machines can be
hard, as it requires some time to identify the needed steps to
install v4l-utils on a new machine, specially because the list
of needed packages should be obtained somewhere. So, let's
document those steps, starting with Debian.

Signed-off-by: Mauro Carvalho Chehab <mchehab@silver.lan>