v4l2-ctl/compliance: add configure option to disable libv4l use
Sometimes you want to compile these two utilities without the libv4l2 wrapper.
This patch adds new configure options to do this:

--disable-v4l2-compliance-libv4l
--disable-v4l2-ctl-libv4l

This is useful when building for an embedded system where you do not want
to use the libv4l2 wrapper library.

The capability to do this was already available, but not hooked up to
the configure system.

Signed-off-by: Hans Verkuil <hansverk@cisco.com>
Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>