qv4l2/v4l2-ctl/v4l2-compliance: fix V4L2_CTRL_TYPE_BITMASK handling
The maximum and default_value field of a bitmask control are defined
as being unsigned instead of signed. This is normally not a problem,
but when you transform a struct v4l2_queryctrl to a struct v4l2_query_ext_ctrl
then you assign an s32 to an s64 and you get sign extension if bit 31 was
set.

So for bitmask controls the s32 should first be cast to a u32 before
assigning it to an s64. The same should be done for struct v4l2_event_ctrl
when the value and/or range changes and a control event is sent.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>