utils: Properly use ENABLE_NLS for locale related code
Various tools in utils/ use ENABLE_NLS to decide whether locale
support is available or not, and only include <locale.h> if ENABLE_NLS
is defined. However, they unconditionally use functions defined in
<locale.h> such as setlocale(), bindtextdomain() or textdomain(),
which causes build failures when the prototypes of such functions are
not available due to <locale.h> not being included.

In order to fix this, we add ENABLE_NLS conditionals around the calls
to these functions.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>