v4l2-compliance: fix buffer flags check
The compliance test verified for a m2m device if the buffer flags of
the original buffer and the processed buffer matched with respect to
time format, timecode and key/b/pframe flags.

However, for compressed formats the timecode flag makes no sense due
to the frame reordering of codecs, and the key/b/pframe flags make
no sense either since the encoder will of course set those flags.

So for compressed formats only check the time format flags.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>