v4l2-ctl: fix cvt/gtf and 'clear timings' bug
When setting new cvt or gtf timings the timings are overwritten due to the
'clear timings' code: since clear_timings is false the driver will get the
old timings and fill in the changed fields. But for CVT and GTF almost all
of those fields where never specified so you end up with wrong timings.

Simplify the code: if timings are specified for CVT or GTF then set clear_timings
to true. And if clear_timings is true, then just use the new timings struct
without the field-by-field copying.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>