v4l: libv4l2subdev: Drop length argument from string conversion functions
v4l2_subdev_string_to_pixelcode() and v4l2_subdev_string_to_field() take a
string and the length of that string as an argument. This has been
motivated by existing usage in the same library. While this works for the
library quite well, it's not a great API.

Instead, drop the length argument and pass a nul terminated string to the
string conversion functions.

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Acked-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>