qv4l2: signal that we don't want GLES
By defining QT_NO_OPENGL_ES_2 we include the non ES headers which
allows us to at least compile qv4l2 on a system where libQt5OpenGL.so
is compiled against GLES.

The application really should test for the required features and
automatically switch to software rendering, but I haven't figured
that out yet. So for now opengl rendering should be disabled
manually.

It would be even better if openGL ES is supported, but that's something
for the future.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>