v4l2-ctl: add streaming support
When using v4l2-ctl to test video capture on devices that do not have a
working video output or just lack the software to easily output captured
video it would be very useful to be able to stream the video to a regular
PC that does have these capabilities.

In the past you would have to capture to file and copy the file over to a
PC, but being able to stream to (or from) a remote host makes life much
easier.

This patch adds --stream-to-host and --stream-from-host options and adds
a simple but extendable protocol for streaming the video.

It does a simple run-length encoding that is optimized for test patterns
(it detects identical consecutive lines and identical consecutive pixels).

For regular video it will likely do a very poor job of compressing the data,
a few percent at best. For test patterns this RLE compresses the frames very,
very effectively.

To view the streamed video a new utility is needed that is still under
construction. This utility will be added to v4l-utils once it is finished.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>