dvbv5-zap: relax the rule of requiring a channel file
Right now, dvbv5-zap will require a fully filled channel file
for most of stuff. The exceptions are when recording all PIDs,
when in monitor or when exiting just after tuning.

Relax the rule to do the reverse: only require fully filled
channel files when recording, when outputing to DVR or when
recording PSI.

That allows using dvbv5-zap as a femon-like tool, tunning into
a transponder and monitoring its signal for a while.

This is useful for debugging purposes, and to monitor the
signals while adjusting the reception.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>