v4l2-compliance: explain the vivid no_error_inj=1 module option
The vivid driver has error injection controls that will cause the
compliance test to fail. To make the tests pass the vivid module needs
to be loaded with the option no_error_inj=1 which disables those controls.

If the compliance test sees lots of failures when testing the vivid
driver, then explain this since it is not obvious if you are not aware
of this.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>