configure.ac: make libjpeg optional
Currently, configure requires libjpeg, except if explicitly
disabled. That's annoying, specially when we're not working with
jpeg webcams (with is the usual usecase - at least for me).

So, make it optional. User will notice anyway if this is missing,
as configure will display the list of options used on the build.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>