libdvbv5: add support for using a remote DVB device
Add a dvbv5-daemon to receive remote DVB device access requests,
and the corresponding logic at the library, via dvb-dev-remote.c.

Please notice that only the libdvbfunctions that start with
"dvb_dev" are supported, with the exception of the dvb_dev_scan().

This is not enough to even tune a device, but support for the
missing parts will be added on further patches.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>