qv4l2/v4l2-ctl: remove obsolete SYCC encoding
The sYCC Y'CbCr encoding is identical to the 601 encoding. Remove
all references to this old name in these applications.

The addition of sYCC was a misunderstanding of the spec that was
unclear about whether sYCC was an extended gamut colorspace or not.
It is not, so that makes it identical to the BT.601 encoding.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>