libdvbv5: prevent craches at the *foreach macro helpers
All those foreach macros assume that not only a struct is
allocated, but also one of their members. This is usually
the case, but it was reported that the logic at channel
scan sometimes get a VCT table, but there are no services
there, causing a crash.

So, add a logic on all those foreach macros to prevent that
such things would ever happen.

Reported-by: Georg Wild <georg.wild@gmx.de>
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>