cec-compliance/follower: fix opcode printing
This contains some small fixes for printing opcodes correctly:

opcode2s should print the message's opcode (based on byte 1), not the Aborted Message
operand for Feature Abort messages (byte 2).

When printing unknown opcode numbers, __u8 must be converted to unsigned
to not print the numbers as characters.

In transmit_timeout the original message struct is kept in order to
print the original message opcode when a Feature Abort occurs.

Signed-off-by: Johan Fjeldtvedt <jaffe1@gmail.com>
Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>