v4l2-ctl: support limited EDID modifications
Add limited support to modify EDIDs. These are primarily various
flags that can modify the behavior, but also the CEC Physical
Address.

Usually used in combination with the built-in EDIDs, but it can
also be used when loading an EDID from a file.

When setting a new EDID these modifiers are printed so it is
easy to see what values these modifiers have.

Also add a new --info-edid option that also prints these modifiers.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>