v4l2-ctl: Add YCbCr 4:4:4 Deep Color, fix HDMI VSDB
The HDMI VSDB modifiers were in two different bytes, not in the
same byte as was mistakenly coded.

This also adds support for the YCbCr 4:4:4 Deep Color bit in that
HDMI VSDB.

The EDIDs were also modified by setting the Graphics bit by default.
If all the CNC bits would be 0, then the HDMI VSDB would be smaller
and no longer contain the CNC bits, making it impossible to modify.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>