cec-follower: check CEC_OP_FEAT_DEV_HAS_SET_OSD_STRING
The follower didn't look at the CEC_OP_FEAT_DEV_HAS_SET_OSD_STRING
feature bit, but cec-compliance did. This resulted in a warning:

        OSD String feature:
                warn: cec-test.cpp(335): The device is in an unsuitable state or cannot display the complete message.

which is due to the fact that the follower implemented this feature
while the compliance test expected it to be absent (since the bit
wasn't set).

The follower should only implement this if the feature bit is set.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>